Don't reduce local variables.
We want to distinguish a term is normal form or not as much as possible
to avoid normalize a normal term again.

We assume terms in environment are normal form.

So reducing a local term to its definition can embeds
a normal term to non-normal-term.

For example, we want to avoid following reduction.

  reduction(beta):
  ((fun n : nat => match n with
		   | 0 => true
		   | S n' => odd n'
		   end) n)
  ->
  match n with
  | 0 => true
  | S n' =>
      (fix even (n : nat) : bool :=
	 match n with
	 | 0 => true
	 | S n'0 => odd n'0
	 end
       with odd (n : nat) : bool :=
	 match n with
	 | 0 => false
	 | S n'0 => even n'0
	 end
       for odd) n'
  end

This commit changes it as follows.

  reduction(beta):
  ((fun n : nat => match n with
		   | 0 => true
		   | S n' => odd n'
		   end) n)
  ->
  match n with
  | 0 => true
  | S n' => odd n'
  end