Don't use Reductionops.beta_applist.
Reductionops.beta_applist is not usable for beta-var
because it apply beta reduction as much as possible.

For example, Reductionops.beta_applist reduces
((fun x y : nat => let z := x + y in Nat.add z) a a)
to
let z := a + a in Nat.add z.

It is problem because beta-var should not
move computation (a + a) from closure invocation time
to closure creation time.

Now,
((fun x y : nat => let z := x + y in Nat.add z) a a)
is reduced to
let x := a in (fun y : nat => let z := x + y in Nat.add z) a).