reduce_eta phase implemented.
This prepends useless multi-functions code generation for
`(fun a => (fix f a := ...) a)`.

It is applied between `replace` and `complete_args`.

Note that eta reduction is not part of conversion rules
because it can change the type of the term as described in
the Coq reference manual.
https://coq.inria.fr/refman/language/core/conversion.html#conversion-rules

But our eta reduction confirms type equality of T1 and T2 in
`(fun (x:T1) => f x)` where `f : T2 -> T3`.
So the types of `(fun (x:T1) => f x)` and `f` are exactly same.